% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate simulated data that meets the parallel trends assumption}
\usage{
generate_data(
  n_obs = 100,
  n_covs_binary = 2,
  n_covs_continuous = 2,
  a_spec,
  y_spec,
  tau = 2,
  potential_outcomes = FALSE
)
}
\arguments{
\item{n_obs}{int. Number of independent observations}

\item{n_covs_binary}{int. Number of binary covariates observed at each period}

\item{n_covs_continuous}{int. Number of continuous covariates observed at each period}

\item{a_spec}{\code{spec} object (as returned by \code{create_spec()}) specifying how the treatments are generated as a function of measured history}

\item{y_spec}{\code{spec} object specifying how the outcomes are generated as a function of measured history}

\item{tau}{int. Number of periods, minus 1. I.e. there are Tt + 1 periods.}

\item{potential_outcomes}{logical. Should outcomes and covariates be generated with exposure set to 0 at all times?}
}
\value{
Data frame with \code{n_obs} \times \code{(tau+1)} rows and \code{n_covs} + 4 columns: \code{uid} is a unique identifier, \code{u} is an 'unmeasured' baseline covariate, \code{a} is a treatment, \code{y} outcomes, and \code{wj} (j=1,...,\code{n_covs}) are covariates.
}
\description{
Generate simulated data that meets the parallel trends assumption
}
