% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_utils.R
\name{create_spec}
\alias{create_spec}
\alias{simulate.spec}
\title{Simulate data from a generalized linear model based on a simulation specification (i.e. a \code{spec} object)}
\usage{
create_spec(family, formula)

\method{simulate}{spec}(object, data, glue_vars = list())
}
\arguments{
\item{family}{a \code{stats::family} object specifying the distribution and link function to simulate from}

\item{formula}{a glue-style formula possibly including time indices. See documentation for glue-style formulas (to be added)}

\item{object}{\code{spec} object, as returned by \code{create_spec()}}

\item{data}{data frame containing the variables referenced in the \code{formula} slot in \code{object}}

\item{params}{optional, numeric vector generalized linear model parameters}

\item{nsim}{integer. number of datasets to return}

\item{seed}{integer. random seed for reproducible results}

\item{...}{values passed to \code{glue_formula()}}
}
\value{
\code{create_spec} returns a \code{spec} object, which is a list containing \code{family} and \code{formula}. \code{simulate.spec} returns a matrix of dimension \code{nrow(data)} \times \code{nsims} containing simulated values.
}
\description{
Simulate data from a generalized linear model based on a simulation specification (i.e. a \code{spec} object)
}
\examples{
a_spec = create_spec(binomial('logit'), '~w1{t}+w2{t}+log(w1{t})+I(w2{t}^2)')
data = data.frame(w11=rnorm(20), w12=rnorm(20))
simulate(a_spec, data=data, t=1)

}
